It "works"
Although it "works" and I can, for a short while, ping flood with a few hundred pps, something is happening to cause the CPU to be halted and no futrher interrupts are acknowledged.

After spending the best part of a week debugging and trying everything, Im no longer making significant progress, so this work is being shelved for now.